test: make a distinction between successful and skipped test
Currently when the test is skipped, it just returns zero and the test
will be assumed to have been successful so the only way to know
which tests were actually skipped is to search the output and inspect
the code.

Change the tests to return -1 in the case that the entire test
is skipped. Some the tests are exercising multiple feature some of which
may not be available and hence can be skipped - I am not changing those
for now, so the rule for the test to be marked as skipped is that
nothing meaningful was actually tested with regards to liburing.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>