test: handle the case when timeout is forced to KILL
Timeout is used to INTerrupt the test if it is still running after
pre-defined time. If the process does not terminate KILL signal is send
after a pre-defined time, but the return status is different in that
case (128+9).
Catch that case and treat is as test failure with the message that the
process had to be killed.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>